+++ public/css/hero-banner.css
@@ -45,12 +45,10 @@
 .sub-category-menu {
     position: absolute;
     left: 100%;
-    top: -20px;
+    top: 0px;
     background: var(--color-white);
     width: 340px;
     max-width: 100%;
-    padding-top: 20px;
-    padding-bottom: 20px;
     opacity: 0;
     visibility: hidden;
     transition: all .2s linear;
@@ -63,7 +61,7 @@
 }
 
 .sub-category-menu li a {
-    padding: 7px 20px;
+    padding: 11px 20px;
     color: var(--color-8A8A8A);
     font-size: 16px;
     display: block;
